perf(home): stream data sections via Suspense, align caching to 5 min
The homepage awaited all backend data (global stats, geo, geo-live, ticker)
plus the nested ML section before flushing any HTML, so TTFB/LCP were gated by
the slowest of ~6 API calls even though the entire above-the-fold hero needs no
remote data.

Split every data-dependent section into its own async Server Component behind a
Suspense boundary with a skeleton fallback:
- HeroTicker (fallback: null, decorative)
- FeaturedParksSlot (SSR seed for the client React Query section)
- GlobalStatsSection (global + platform stats from one getGlobalStats fetch)
- MLStatsSection (now wrapped in Suspense)
- LiveActivitySection (geo + geo-live merge)

The hero shell now streams immediately; sections stream in as their fetches
resolve. getGeoStructure is request-deduped between the featured-parks and
live-activity sections, so no extra round-trip.

Caching: page revalidate 60 -> 300 to match the 5 min fetch-level data cache
used by all live endpoints (ML stays at 1800). The route is dynamically rendered
(the [locale] layout reads cookies()), so streaming + the 5 min fetch data cache
together give fast first paint with ~5 min-fresh data.
