+++ app/[locale]/page.tsx
@@ -1,35 +1,18 @@
+import { Suspense } from 'react';
 import { getTranslations, setRequestLocale } from 'next-intl/server';
 import { generateAlternateLanguages, locales } from '@/i18n/config';
 import { buildOpenGraphMetadata } from '@/lib/utils/metadata';
 import { Link } from '@/i18n/navigation';
 import { GLOSSARY_SEGMENTS } from '@/lib/glossary/translations';
 import type { Locale } from '@/i18n/config';
-import {
-  Clock,
-  TrendingUp,
-  ChevronRight,
-  Map as MapIcon,
-  BookOpen,
-  Tag,
-  BarChart3,
-  Database,
-  Globe,
-  Sparkles,
-} from 'lucide-react';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { Clock, TrendingUp, Map as MapIcon, BookOpen, Tag, Sparkles } from 'lucide-react';
 import { Button } from '@/components/ui/button';
-import { getGlobalStats, getGeoLiveStats, getTickerData } from '@/lib/api/analytics';
-import { catchNonFatal } from '@/lib/api/client';
-import { LiveWaitTicker } from '@/components/home/live-wait-ticker';
-import { getGeoStructure } from '@/lib/api/discovery';
 
 import nextDynamic from 'next/dynamic';
 import { HeroBackground } from '@/components/layout/hero-background';
 import { HERO_IMAGES } from '@/lib/hero-images';
 import { HomepageFAQStructuredData } from '@/components/seo/homepage-faq-structured-data';
-import { OpenStatusProgress } from '@/components/common/open-status-progress';
 import { GlassCard } from '@/components/common/glass-card';
-import { translateGeoSlug } from '@/lib/utils/geo-translate';
 
 const LocationBanner = nextDynamic(
   () => import('@/components/common/location-banner').then((m) => ({ default: m.LocationBanner })),
@@ -54,27 +37,32 @@ const NearbyParksCard = nextDynamic(
     ssr: true,
   }
 );
-import { StatsCard } from '@/components/common/stats-card';
-import { convertApiUrlToFrontendUrl } from '@/lib/utils/url-utils';
-import { CompactNumberWithTooltip } from '@/components/common/compact-number-with-tooltip';
 import { AnnounceSection } from '@/components/home/announce-section';
 import { MLStatsSection } from '@/components/home/ml-stats-section';
-import { ParkCard } from '@/components/parks/park-card';
-import { AttractionCard } from '@/components/parks/attraction-card';
-import { getParkBackgroundImage, getAttractionBackgroundImage } from '@/lib/utils/park-assets';
 import { HeroImageInfo } from '@/components/layout/hero-image-info';
 import { HERO_IMAGE_META } from '@/lib/hero-images-meta';
 import { HeroWithNearby } from '@/components/home/hero-with-nearby';
 import { HeroSearchInput } from '@/components/search/hero-search-input';
-import { FeaturedParksSectionClient } from '@/components/home/featured-parks-section-client';
-import { extractFeaturedParks } from '@/components/home/featured-parks-section';
+import { HeroTicker } from '@/components/home/hero-ticker';
+import { FeaturedParksSlot } from '@/components/home/featured-parks-slot';
+import { GlobalStatsSection } from '@/components/home/global-stats-section';
+import { LiveActivitySection } from '@/components/home/live-activity-section';
+import {
+  GlobalStatsSkeleton,
+  FeaturedParksSkeleton,
+  MLStatsSkeleton,
+  LiveActivitySkeleton,
+} from '@/components/home/home-skeletons';
 
 import { getOgImageUrl } from '@/lib/utils/og-image';
 import { GlossaryInject } from '@/components/glossary/glossary-inject';
 
 import type { Metadata } from 'next';
 
-export const revalidate = 60;
+// ~5 min ISR window for the static shell. Each data section sets its own
+// revalidate on its fetch (analytics/geo: 300s, ML: 1800s) and streams in via
+// its own Suspense boundary, so a slow/stale endpoint never blocks first paint.
+export const revalidate = 300;
 
 export async function generateStaticParams() {
   return locales.map((locale) => ({ locale }));
@@ -116,38 +104,13 @@ export default async function HomePage({ params }: HomePageProps) {
   setRequestLocale(locale);
   const glossaryPath = '/' + GLOSSARY_SEGMENTS[locale as Locale];
 
-  // Kick off the backend data fetch FIRST so the round-trips overlap with the (cheap, but
-  // previously 6× sequential) translation loading below instead of waiting behind it.
-  const dataPromise = Promise.all([
-    catchNonFatal(getGlobalStats()),
-    catchNonFatal(getGeoStructure(300)),
-    catchNonFatal(getGeoLiveStats()),
-    catchNonFatal(getTickerData()),
-  ]);
-
-  const [t, tHome, tParks, tCommon, tGeo, tExplore] = await Promise.all([
-    getTranslations('stats'),
-    getTranslations('home'),
-    getTranslations('parks'),
-    getTranslations('common'),
-    getTranslations('geo'),
-    getTranslations('explore'),
-  ]);
+  // Only the static, above-the-fold shell needs translations up front. Every
+  // data-dependent section fetches its own data (and translations) inside a
+  // Suspense boundary below, so the hero renders/streams without waiting on the API.
+  const [tHome, tParks] = await Promise.all([getTranslations('home'), getTranslations('parks')]);
 
   // eslint-disable-next-line react-hooks/purity -- intentional per-request randomization
   const randomHeroImage = HERO_IMAGES[Math.floor(Math.random() * HERO_IMAGES.length)];
-
-  // ML dashboard is fetched inside MLStatsSection to keep Suspense boundaries clean.
-  const [stats, geoData, liveStats, tickerData] = await dataPromise;
-
-  const continents =
-    geoData?.continents.map((continent) => {
-      const liveContinent = liveStats?.continents.find((c) => c.slug === continent.slug);
-      return {
-        ...continent,
-        openParkCount: liveContinent?.openParkCount ?? continent.openParkCount,
-      };
-    }) || [];
 
   return (
     <div className="flex flex-col">
@@ -216,12 +179,10 @@ export default async function HomePage({ params }: HomePageProps) {
           <HeroImageInfo meta={HERO_IMAGE_META[randomHeroImage]} />
         )}
 
-        {/* Live wait times ticker — hidden on phones (decorative, absolute → no layout impact) */}
-        {tickerData && tickerData.items.length > 0 && (
-          <div className="absolute right-0 bottom-0 left-0 hidden md:block">
-            <LiveWaitTicker initialItems={tickerData.items} />
-          </div>
-        )}
+        {/* Live wait times ticker — streamed in; never blocks the hero (decorative, absolute) */}
+        <Suspense fallback={null}>
+          <HeroTicker />
+        </Suspense>
       </section>
 
       {/* Announcement Section */}
@@ -240,335 +201,25 @@ export default async function HomePage({ params }: HomePageProps) {
       {/* Favorites Section */}
       <FavoritesSection />
 
-      {/* Featured Parks – locale-aware, direct park links for SEO */}
-      <FeaturedParksSectionClient
-        headingText={tHome('sections.featuredParks')}
-        introText={tHome('sections.featuredParksIntro')}
-        ctaText={tHome('hero.cta')}
-        initialParks={extractFeaturedParks(geoData, locale as string)}
-      />
+      {/* Featured Parks – locale-aware, direct park links for SEO (SSR seed + client live data) */}
+      <Suspense fallback={<FeaturedParksSkeleton />}>
+        <FeaturedParksSlot locale={locale} />
+      </Suspense>
 
-      {/* Global Stats */}
-      {stats && (
-        <section className="bg-muted/30 px-4 py-12">
-          <div className="container mx-auto">
-            <div className="mb-2 flex items-center gap-2">
-              <BarChart3 className="text-primary h-5 w-5" />
-              <h2 className="text-xl font-bold">{t('globalStats')}</h2>
-            </div>
-            <p className="text-muted-foreground mb-8 text-sm">{t('globalStatsIntro')}</p>
+      {/* Global Stats + Platform Statistics (single getGlobalStats fetch) */}
+      <Suspense fallback={<GlobalStatsSkeleton />}>
+        <GlobalStatsSection />
+      </Suspense>
 
-            {/* Grid Layout: First row - 2 static cards */}
-            <div className="mb-4 grid gap-4 sm:grid-cols-2">
-              {/* Open Parks */}
-              <StatsCard
-                title={t('openParks')}
-                value={stats.counts.openParks}
-                description={
-                  <>
-                    {tCommon('of')} {stats.counts.parks} {tCommon('total')}
-                  </>
-                }
-              />
-
-              {/* Total Attractions */}
-              <StatsCard
-                title={t('totalAttractions')}
-                value={stats.counts.attractions.toLocaleString()}
-                description={
-                  <>
-                    {stats.counts.openAttractions.toLocaleString()} {tCommon('operating')}
-                  </>
-                }
-              />
-            </div>
-
-            {/* Grid Layout: Second row - Parks */}
-            <div className="mb-3 grid gap-4 sm:grid-cols-2">
-              {stats.mostCrowdedPark && (
-                <div className="grid [grid-template-rows:auto_auto_1fr_auto] gap-4">
-                  <h3 className="text-muted-foreground text-sm font-medium">{t('mostCrowded')}</h3>
-                  <ParkCard
-                    name={stats.mostCrowdedPark.name}
-                    slug={stats.mostCrowdedPark.slug}
-                    parkId={stats.mostCrowdedPark.id}
-                    city={stats.mostCrowdedPark.city}
-                    country={translateGeoSlug(
-                      tGeo,
-                      'countries',
-                      stats.mostCrowdedPark.countrySlug,
-                      stats.mostCrowdedPark.country
-                    )}
-                    href={convertApiUrlToFrontendUrl(stats.mostCrowdedPark.url) as '/'}
-                    backgroundImage={getParkBackgroundImage(stats.mostCrowdedPark.slug)}
-                    status="OPERATING"
-                    timezone={stats.mostCrowdedPark.timezone}
-                    crowdLevel={stats.mostCrowdedPark.crowdLevel ?? undefined}
-                    averageWaitTime={stats.mostCrowdedPark.averageWaitTime ?? undefined}
-                    operatingAttractions={stats.mostCrowdedPark.operatingAttractions}
-                    totalAttractions={stats.mostCrowdedPark.totalAttractions}
-                  />
-                </div>
-              )}
-              {stats.leastCrowdedPark && (
-                <div className="grid [grid-template-rows:auto_auto_1fr_auto] gap-4">
-                  <h3 className="text-muted-foreground text-sm font-medium">{t('leastCrowded')}</h3>
-                  <ParkCard
-                    name={stats.leastCrowdedPark.name}
-                    slug={stats.leastCrowdedPark.slug}
-                    parkId={stats.leastCrowdedPark.id}
-                    city={stats.leastCrowdedPark.city}
-                    country={translateGeoSlug(
-                      tGeo,
-                      'countries',
-                      stats.leastCrowdedPark.countrySlug,
-                      stats.leastCrowdedPark.country
-                    )}
-                    href={convertApiUrlToFrontendUrl(stats.leastCrowdedPark.url) as '/'}
-                    backgroundImage={getParkBackgroundImage(stats.leastCrowdedPark.slug)}
-                    status="OPERATING"
-                    timezone={stats.leastCrowdedPark.timezone}
-                    crowdLevel={stats.leastCrowdedPark.crowdLevel ?? undefined}
-                    averageWaitTime={stats.leastCrowdedPark.averageWaitTime ?? undefined}
-                    operatingAttractions={stats.leastCrowdedPark.operatingAttractions}
-                    totalAttractions={stats.leastCrowdedPark.totalAttractions}
-                  />
-                </div>
-              )}
-            </div>
-
-            {/* Grid Layout: Third row - Attractions */}
-            <div className="grid gap-4 sm:grid-cols-2">
-              {stats.longestWaitRide && (
-                <div className="grid [grid-template-rows:auto_auto_1fr_auto] gap-4">
-                  <h3 className="text-muted-foreground text-sm font-medium">{t('longestWait')}</h3>
-                  <AttractionCard
-                    parkStatus="OPERATING"
-                    showParkName
-                    backgroundImage={
-                      getAttractionBackgroundImage(
-                        stats.longestWaitRide.parkSlug,
-                        stats.longestWaitRide.slug
-                      ) ?? getParkBackgroundImage(stats.longestWaitRide.parkSlug)
-                    }
-                    attraction={{
-                      id: stats.longestWaitRide.id,
-                      name: stats.longestWaitRide.name,
-                      slug: stats.longestWaitRide.slug,
-                      url: convertApiUrlToFrontendUrl(stats.longestWaitRide.url),
-                      latitude: null,
-                      longitude: null,
-                      crowdLevel: stats.longestWaitRide.crowdLevel ?? undefined,
-                      queues: [
-                        {
-                          queueType: 'STANDBY',
-                          waitTime: stats.longestWaitRide.waitTime,
-                          status: 'OPERATING',
-                        },
-                      ],
-                      statistics: stats.longestWaitRide.sparkline.length
-                        ? {
-                            avgWaitToday: stats.longestWaitRide.avgWaitToday,
-                            minWaitToday: stats.longestWaitRide.minWaitToday,
-                            peakWaitToday: stats.longestWaitRide.peakWaitToday,
-                            peakWaitTimestamp: stats.longestWaitRide.peakWaitTimestamp,
-                            typicalWaitThisHour: stats.longestWaitRide.typicalWaitThisHour,
-                            percentile95ThisHour: null,
-                            currentVsTypical: stats.longestWaitRide.currentVsTypical,
-                            dataPoints: stats.longestWaitRide.sparkline.length,
-                            history: stats.longestWaitRide.sparkline,
-                            timestamp: new Date().toISOString(),
-                          }
-                        : undefined,
-                      park: {
-                        id: '',
-                        name: stats.longestWaitRide.parkName,
-                        slug: stats.longestWaitRide.parkSlug,
-                        timezone: stats.longestWaitRide.parkTimezone,
-                        continent: null,
-                        country: stats.longestWaitRide.parkCountrySlug,
-                        city: stats.longestWaitRide.parkCity,
-                      },
-                    }}
-                  />
-                </div>
-              )}
-              {stats.shortestWaitRide && (
-                <div className="grid [grid-template-rows:auto_auto_1fr_auto] gap-4">
-                  <h3 className="text-muted-foreground text-sm font-medium">{t('shortestWait')}</h3>
-                  <AttractionCard
-                    parkStatus="OPERATING"
-                    showParkName
-                    backgroundImage={
-                      getAttractionBackgroundImage(
-                        stats.shortestWaitRide.parkSlug,
-                        stats.shortestWaitRide.slug
-                      ) ?? getParkBackgroundImage(stats.shortestWaitRide.parkSlug)
-                    }
-                    attraction={{
-                      id: stats.shortestWaitRide.id,
-                      name: stats.shortestWaitRide.name,
-                      slug: stats.shortestWaitRide.slug,
-                      url: convertApiUrlToFrontendUrl(stats.shortestWaitRide.url),
-                      latitude: null,
-                      longitude: null,
-                      crowdLevel: stats.shortestWaitRide.crowdLevel ?? undefined,
-                      queues: [
-                        {
-                          queueType: 'STANDBY',
-                          waitTime: stats.shortestWaitRide.waitTime,
-                          status: 'OPERATING',
-                        },
-                      ],
-                      statistics: stats.shortestWaitRide.sparkline.length
-                        ? {
-                            avgWaitToday: stats.shortestWaitRide.avgWaitToday,
-                            minWaitToday: stats.shortestWaitRide.minWaitToday,
-                            peakWaitToday: stats.shortestWaitRide.peakWaitToday,
-                            peakWaitTimestamp: stats.shortestWaitRide.peakWaitTimestamp,
-                            typicalWaitThisHour: stats.shortestWaitRide.typicalWaitThisHour,
-                            percentile95ThisHour: null,
-                            currentVsTypical: stats.shortestWaitRide.currentVsTypical,
-                            dataPoints: stats.shortestWaitRide.sparkline.length,
-                            history: stats.shortestWaitRide.sparkline,
-                            timestamp: new Date().toISOString(),
-                          }
-                        : undefined,
-                      park: {
-                        id: '',
-                        name: stats.shortestWaitRide.parkName,
-                        slug: stats.shortestWaitRide.parkSlug,
-                        timezone: stats.shortestWaitRide.parkTimezone,
-                        continent: null,
-                        country: stats.shortestWaitRide.parkCountrySlug,
-                        city: stats.shortestWaitRide.parkCity,
-                      },
-                    }}
-                  />
-                </div>
-              )}
-            </div>
-          </div>
-        </section>
-      )}
-
-      {/* Platform Statistics */}
-      {stats && (
-        <section className="px-4 py-12">
-          <div className="container mx-auto">
-            <div className="mb-2 flex items-center gap-2">
-              <Database className="text-primary h-5 w-5" />
-              <h2 className="text-xl font-bold">{t('platformStats')}</h2>
-            </div>
-            <p className="text-muted-foreground mb-8 text-sm">{t('platformStatsDescription')}</p>
-
-            <div className="grid gap-4 sm:grid-cols-2 lg:grid-cols-3">
-              {/* Total Wait Time */}
-              {stats.counts.totalWaitTime != null && (
-                <StatsCard
-                  title={t('totalWaitTime')}
-                  value={stats.counts.totalWaitTime.toLocaleString()}
-                  description={
-                    <>
-                      {tCommon('minutes')} · ~{Math.round(stats.counts.totalWaitTime / 60)}{' '}
-                      {tCommon('hours')}
-                    </>
-                  }
-                />
-              )}
-
-              {/* Queue Data Records */}
-              <StatsCard
-                title={t('dataPoints')}
-                value={<CompactNumberWithTooltip value={stats.counts.queueDataRecords} />}
-                description={t('queueDataRecords')}
-              />
-
-              {/* Shows & Restaurants */}
-              <Card>
-                <CardHeader className="pb-2">
-                  <CardTitle className="text-muted-foreground text-sm font-medium">
-                    {t('alsoTracking')}
-                  </CardTitle>
-                </CardHeader>
-                <CardContent>
-                  <div className="mb-1 flex items-baseline gap-2">
-                    <span className="text-2xl font-bold">{stats.counts.shows}</span>
-                    <span className="text-muted-foreground text-sm">{tCommon('shows')}</span>
-                  </div>
-                  <div className="flex items-baseline gap-2">
-                    <span className="text-2xl font-bold">{stats.counts.restaurants}</span>
-                    <span className="text-muted-foreground text-sm">{tCommon('restaurants')}</span>
-                  </div>
-                </CardContent>
-              </Card>
-            </div>
-          </div>
-        </section>
-      )}
       {/* ML / AI Stats */}
-      <MLStatsSection />
+      <Suspense fallback={<MLStatsSkeleton />}>
+        <MLStatsSection />
+      </Suspense>
 
       {/* Live Activity - Parks Open Now */}
-      {continents.length > 0 && (
-        <section className="px-4 py-12">
-          <div className="container mx-auto">
-            <div className="mb-2 flex items-center gap-2">
-              <Globe className="text-primary h-5 w-5" />
-              <h2 className="text-xl font-bold">{tHome('sections.liveNow')}</h2>
-            </div>
-            <p className="text-muted-foreground mb-8 text-sm">{tHome('sections.liveNowIntro')}</p>
-            <div className="grid gap-4 sm:grid-cols-2 lg:grid-cols-3">
-              {continents.map((continent) => {
-                const continentName = translateGeoSlug(
-                  tGeo,
-                  'continents',
-                  continent.slug,
-                  continent.name
-                );
-
-                return (
-                  <Link
-                    key={continent.slug}
-                    href={`/parks/${continent.slug}`}
-                    prefetch={false}
-                    className="group block"
-                  >
-                    <Card className="bg-muted/50 hover:bg-muted/70 border-border hover:border-primary/50 h-full transition-all duration-200 hover:scale-[1.02] hover:shadow-lg">
-                      <CardHeader className="pb-2">
-                        <div className="flex items-center justify-between">
-                          <CardTitle className="text-lg font-medium">{continentName}</CardTitle>
-                          <ChevronRight className="text-muted-foreground group-hover:text-primary h-4 w-4 transition-colors" />
-                        </div>
-                        <div className="text-muted-foreground text-xs">
-                          {continent.countryCount}{' '}
-                          {tExplore('stats.country', { count: continent.countryCount })}
-                        </div>
-                      </CardHeader>
-                      <CardContent>
-                        <div className="mb-2 flex items-baseline gap-2">
-                          <span className="bg-gradient-to-r from-green-600 to-emerald-500 bg-clip-text text-3xl font-bold text-transparent">
-                            {continent.openParkCount}
-                          </span>
-                          <span className="text-muted-foreground text-sm">
-                            / {continent.parkCount}
-                          </span>
-                        </div>
-                        {/* Progress bar */}
-                        <OpenStatusProgress
-                          openCount={continent.openParkCount}
-                          totalCount={continent.parkCount}
-                          showLabel={false}
-                        />
-                      </CardContent>
-                    </Card>
-                  </Link>
-                );
-              })}
-            </div>
-          </div>
-        </section>
-      )}
+      <Suspense fallback={<LiveActivitySkeleton />}>
+        <LiveActivitySection />
+      </Suspense>
 
       {/* Features Section */}
       <section className="bg-muted/30 px-4 py-16">
